+++ changelog.md
@@ -2,6 +2,7 @@
 
 - It is now possible to specify as many submenus in the `menus` array as you want
 - Bugfix: The CKEditor in WYSIWYG fields no longer jumps around and removes text selection on blur
+- Bugfix: Soft deleted values from a related table no longer get included in relationship columns
 - Bugfix: Relationship where clauses now work when you pre-specify the table name
 - Bugfix: Time fields weren't saving properly